requirements.txt
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
qdrant-client>=1.12.0
groq>=0.11.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pytest>=8.0.0
